+++ common.cpp
@@ -1,8 +1,8 @@
 #include "common.h"
 
 void rev_comp_inplace(std::string & s) {
-    std::string tmp = s; int l = s.length();
-    for(int i = 0; i < l; ++i) {
+    std::string tmp = s; size_t l = s.length();
+    for(size_t i = 0; i < l; ++i) {
         s[i] = COMPLEMENT[(int)tmp[l-i-1]];
     }
 }
